Fix routes for admin

+++ routes/web.php
@@ -26,12 +26,12 @@ Route::middleware('auth')->group(function () {
 Route::middleware('IsUserAdmin')->group(function () {
     Route::get('/admin', [AdminDashboardController::class, 'index'])->name('admin');
 
-    Route::get('/admin/users', [UserController::class, 'index'])->name('admin.users');
-    Route::get('/admin/users/create', [UserController::class, 'create'])->name('admin.users.create');
-    Route::post('/admin/users', [UserController::class, 'store'])->name('admin.users.store');
-    Route::get('/admin/users/{user}/edit', [UserController::class, 'edit'])->name('admin.users.edit');
-    Route::patch('/admin/users/{user}', [UserController::class, 'update'])->name('admin.users.update');
-    Route::delete('/admin/users/{user}', [UserController::class, 'destroy'])->name('admin.users.destroy');
+    Route::get('/admin/users', [\App\Http\Controllers\Admin\UserController::class, 'index'])->name('admin.users');
+    Route::get('/admin/users/create', [\App\Http\Controllers\Admin\UserController::class, 'create'])->name('admin.users.create');
+    Route::post('/admin/users', [\App\Http\Controllers\Admin\UserController::class, 'store'])->name('admin.users.store');
+    Route::get('/admin/users/{user}/edit', [\App\Http\Controllers\Admin\UserController::class, 'edit'])->name('admin.users.edit');
+    Route::patch('/admin/users/{user}', [\App\Http\Controllers\Admin\UserController::class, 'update'])->name('admin.users.update');
+    Route::delete('/admin/users/{user}', [\App\Http\Controllers\Admin\UserController::class, 'destroy'])->name('admin.users.destroy');
 });
 
 Route::get('/user/{id}', [UserController::class, 'getUser'])->name('getUser');
@@ -64,5 +64,7 @@ Route::post('/invite', [InviteController::class, 'store'])->name('store')->middl
 Route::post('/invite/{verification_token}', [InviteController::class, 'update'])->middleware('auth.basic');
 Route::get('/invites', [InviteController::class, 'index'])->middleware('auth.basic');
 Route::delete('/invite/{invite}', [InviteController::class, 'destroy'])->middleware('auth.basic');
+
+
 
 require __DIR__.'/auth.php';
